Book adding features working, Delete not working
This is a progress update of the page, everything but the book Read and
the book delete is working. The page keeps track of the books a user
reads, and allows the user to add more. The focus learning of this
lesson was a refresher on javascript, as well as a focus on OOP. Mostly
for me though, it was a review of syntax in .js.