Delete and Read buttons on the book cards now work
Added the functionality to the buttons to make them delete and be able
to change the book from read to unread. The color of the button simply
changes from red to green depending on if the user has read or unread
the book.

Still need to fix the flexbox on the modal popup, as well as more CSS
styling couldn't hurt.